+++ app/controllers/boats_controller.ts
@@ -150,7 +150,7 @@ export default class BoatsController {
         this.sheetService.listForBoat(user, boat),
         this.mediaService.listForEntity('boat', boat.id),
         this.boatService.getPositionHistory(boat.id),
-        this.aiAnalysisService.getLatestBoatSuggestions(user.id, boat.id),
+        this.aiAnalysisService.getLatestBoatSuggestions(user.id, boat.id, user.organization.id),
         bouncer.with(BoatPolicy).allows('edit', boat),
         this.documentService.listForBoat(user, boat),
       ])